# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UAC API Documentation",
    "version": "1.0.0",
    "description": "用户认证和授权系统 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SSOConfig": {
        "type": "object",
        "required": [
          "protocol",
          "redirect_uri",
          "salt"
        ],
        "properties": {
          "protocol": {
            "type": "string",
            "enum": [
              "SAML",
              "CAS",
              "OIDC",
              "OAuth"
            ],
            "description": "SSO使用的协议",
            "example": "OIDC"
          },
          "redirect_uri": {
            "type": "string",
            "format": "uri",
            "description": "SSO回调地址",
            "example": "https://hrms.example.com/auth/callback"
          },
          "salt": {
            "type": "string",
            "description": "SSO签名盐值，用于JWT签名",
            "example": "sso-salt-123"
          },
          "secret": {
            "type": "string",
            "description": "基于currenttime、salt，使用 bcrypt 生成的Hash值",
            "example": "$2a$10$xxxxx"
          },
          "currentTimestample": {
            "type": "integer",
            "description": "当前时间戳， 用于生成secret",
            "example": 17198592000
          },
          "additional_params": {
            "type": "object",
            "description": "其他SSO协议特定的参数",
            "additionalProperties": true,
            "example": {
              "client_id": "hrms-client",
              "client_secret": "your-client-secret",
              "issuer": "https://sso.example.com"
            }
          }
        }
      },
      "APIConnectConfig": {
        "type": "object",
        "required": [
          "salt"
        ],
        "properties": {
          "app_secret": {
            "type": "string",
            "description": "应用API私钥（由服务端根据 application_id 和 salt 生成）",
            "example": "wLTAwMDAtMDAwMDAwMDAwMDAxIiwidXNlcm5hbWUiOiJhZG1pbiIsImlhdCI6MTc0OTE4MTAyMCwiZXhwIjoxNzQ5MjY3NDIwfQ.VOfXDBi5DeWGTsAMzRmBNfP4AikJhT6WevpupizBrm4"
          },
          "salt": {
            "type": "string",
            "description": "签名盐值",
            "example": "random-salt-456"
          }
        }
      },
      "APIDataScope": {
        "type": "object",
        "description": "API数据权限范围配置",
        "additionalProperties": {
          "type": "string",
          "enum": [
            "all",
            "department",
            "self"
          ],
          "description": "数据权限范围"
        },
        "example": {
          "user:read": "department",
          "user:write": "self",
          "department:read": "all"
        }
      },
      "Application": {
        "type": "object",
        "properties": {
          "application_id": {
            "type": "string",
            "format": "uuid",
            "description": "应用ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "description": "应用名称",
            "example": "人力资源管理系统"
          },
          "code": {
            "type": "string",
            "description": "应用编码",
            "example": "hrms"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ],
            "description": "应用状态",
            "example": "ACTIVE"
          },
          "sso_enabled": {
            "type": "boolean",
            "description": "是否启用SSO",
            "example": true
          },
          "sso_config": {
            "$ref": "#/components/schemas/SSOConfig"
          },
          "api_enabled": {
            "type": "boolean",
            "description": "是否启用API服务",
            "example": true
          },
          "api_connect_config": {
            "$ref": "#/components/schemas/APIConnectConfig"
          },
          "api_data_scope": {
            "$ref": "#/components/schemas/APIDataScope"
          },
          "description": {
            "type": "string",
            "description": "应用描述",
            "example": "公司人力资源管理系统"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间",
            "example": "2024-03-21T10:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间",
            "example": "2024-03-21T10:00:00.000Z"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "删除时间",
            "example": null
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/v1/applications": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "创建应用",
        "description": "创建一个新的应用",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "code"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "应用名称",
                    "example": "人力资源管理系统"
                  },
                  "code": {
                    "type": "string",
                    "description": "应用编码",
                    "example": "hrms"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "DISABLED"
                    ],
                    "description": "应用状态",
                    "example": "ACTIVE"
                  },
                  "sso_enabled": {
                    "type": "boolean",
                    "description": "是否启用SSO",
                    "example": true
                  },
                  "sso_config": {
                    "$ref": "#/components/schemas/SSOConfig"
                  },
                  "api_enabled": {
                    "type": "boolean",
                    "description": "是否启用API服务",
                    "example": true
                  },
                  "api_connect_config": {
                    "$ref": "#/components/schemas/APIConnectConfig"
                  },
                  "api_data_scope": {
                    "$ref": "#/components/schemas/APIDataScope"
                  },
                  "description": {
                    "type": "string",
                    "description": "应用描述",
                    "example": "公司人力资源管理系统"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "创建成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Application"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "获取应用列表",
        "description": "获取应用列表，支持分页和筛选。当 size 参数为 -1 时，返回所有记录不分页。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "description": "页码（当 size 不为 -1 时有效）"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": -1,
              "maximum": 100
            },
            "description": "每页数量，设置为 -1 时返回所有记录不分页"
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "应用名称（支持模糊匹配）"
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "应用编码（支持模糊匹配）"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "DISABLED"
              ]
            },
            "description": "应用状态"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "description": "总记录数（当 size 为 -1 时等于 items 的长度）",
                          "example": 100
                        },
                        "page": {
                          "type": "integer",
                          "description": "当前页码（当 size 为 -1 时固定为 1）",
                          "example": 1
                        },
                        "size": {
                          "type": "integer",
                          "description": "每页数量（当 size 为 -1 时等于总记录数）",
                          "example": 10
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Application"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/v1/applications/{id}": {
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "获取应用详情",
        "description": "根据ID获取应用详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "应用ID"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "获取成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Application"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "应用不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Applications"
        ],
        "summary": "更新应用",
        "description": "更新应用信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "应用ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "应用名称",
                    "example": "人力资源管理系统"
                  },
                  "code": {
                    "type": "string",
                    "description": "应用编码",
                    "example": "hrms"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "DISABLED"
                    ],
                    "description": "应用状态",
                    "example": "ACTIVE"
                  },
                  "sso_enabled": {
                    "type": "boolean",
                    "description": "是否启用SSO",
                    "example": true
                  },
                  "sso_config": {
                    "$ref": "#/components/schemas/SSOConfig"
                  },
                  "api_enabled": {
                    "type": "boolean",
                    "description": "是否启用API服务",
                    "example": true
                  },
                  "api_connect_config": {
                    "$ref": "#/components/schemas/APIConnectConfig"
                  },
                  "api_data_scope": {
                    "$ref": "#/components/schemas/APIDataScope"
                  },
                  "description": {
                    "type": "string",
                    "description": "应用描述",
                    "example": "公司人力资源管理系统"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "更新成功"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Application"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "应用不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Applications"
        ],
        "summary": "删除应用",
        "description": "删除指定应用",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "应用ID"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "删除成功"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "应用不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{id}/generate-secret": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "生成应用密钥",
        "description": "根据应用ID和salt生成app_secret，并保存到应用的api_connect_config中",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "应用ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "salt"
                ],
                "properties": {
                  "salt": {
                    "type": "string",
                    "description": "签名盐值",
                    "example": "random-salt-123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "生成成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "app_secret": {
                          "type": "string",
                          "description": "应用密钥",
                          "example": "a1b2c3d4e5f6g7h8i9j0..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "应用不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "服务器错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}